+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/mcpdirectstudioapp/App.kt
@@ -15,8 +15,7 @@ import ai.mcpdirect.studio.app.setting.SettingsScreen
 import ai.mcpdirect.studio.app.team.MCPTeamScreen
 import ai.mcpdirect.studio.app.team.MCPTeamToolMakerScreen
 import ai.mcpdirect.studio.app.team.MCPTeamToolMakerTemplateScreen
-import ai.mcpdirect.studio.app.theme.blue.BlueTheme
-import ai.mcpdirect.studio.app.theme.purple.PurpleTheme
+import ai.mcpdirect.studio.app.theme.AppTheme
 import ai.mcpdirect.studio.app.tips.QuickStartScreen
 import ai.mcpdirect.studio.app.tips.TipsScreen
 import ai.mcpdirect.studio.app.tool.MCPToolsScreen
@@ -63,7 +62,7 @@ fun App(
 //            }
 //        }
 //    }
-    BlueTheme (
+    AppTheme (
         darkTheme = darkTheme,
         typography = typography
     ) {